% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anzia2012}
\alias{anzia2012}
\title{Dataset from Anzia (2012)}
\format{
A panel data in \code{tbl} and \code{data.frame} format
  with 6,377 rows and 11 variables:
 \describe{
   \item{district}{district id}
   \item{year}{time index}
   \item{lnavgsalary_cpi}{main outcome variable}
   \item{oncycle}{main treatment variable}
 }
}
\usage{
data(anzia2012)
}
\description{
Dataset from Anzia (2012)
}
\references{
Anzia, Sarah F.
 "The Election Timing Effect: Evidence from a Policy Intervention in Texas."
 \emph{Quarterly Journal of Political Science} 7.3 (2012):209-248.
 \doi{10.1561/100.00011056}
}
\keyword{dataset}
