% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{compute_sdid}
\alias{compute_sdid}
\title{Estimate sDID}
\usage{
compute_sdid(dat_panel, outcome, treatment, id_time_use, id_subj_use,
  time_weight, min_time = 3, lead)
}
\description{
Estimate sDID
}
\keyword{internal}
