% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad-functions.R
\name{create_Gmat}
\alias{create_Gmat}
\title{Create a G matrix}
\usage{
create_Gmat(dat_panel, treatment)
}
\arguments{
\item{dat_panel}{A class of \code{panelr} object.}
}
\value{
A matrix of the treatment assignment timing.
}
\description{
Create a G matrix
}
\keyword{internal}
