% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric.R
\name{cugmm_loss_parametric_resid}
\alias{cugmm_loss_parametric_resid}
\title{gmm loss function}
\usage{
cugmm_loss_parametric_resid(par, dat, id_subject, k, p = 1)
}
\arguments{
\item{par}{parameter vector.}

\item{dat}{list of data. List of outputs from \code{getX} function.}

\item{id_subject}{vector of subject id.}

\item{k}{number of outcome types (models)}

\item{p}{number of variables}
}
\description{
gmm loss function
}
\keyword{internal}
