% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{cugmm_loss_rcs_init}
\alias{cugmm_loss_rcs_init}
\title{CU-GMM loss function for RCS data}
\usage{
cugmm_loss_rcs_init(par, dat)
}
\arguments{
\item{par}{parameter.}

\item{dat}{a list of data consists of y_resid and d_reisd.}
}
\description{
CU-GMM loss function for RCS data
}
\keyword{internal}
