% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric.R
\name{cugmm_var_parametric}
\alias{cugmm_var_parametric}
\title{compute asymptotic GMM variance}
\usage{
cugmm_var_parametric(par, dat, id_subject)
}
\description{
compute asymptotic GMM variance
}
\keyword{internal}
