% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric.R
\name{cugmm_var_parametric_resid}
\alias{cugmm_var_parametric_resid}
\title{compute asymptotic GMM variance}
\usage{
cugmm_var_parametric_resid(par, dat, id_subject, p = 1)
}
\description{
compute asymptotic GMM variance
}
\keyword{internal}
