% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{cugmm_var_rcs}
\alias{cugmm_var_rcs}
\title{CU-GMM variance function for RCS data}
\usage{
cugmm_var_rcs(par, dat)
}
\arguments{
\item{par}{parameter.}

\item{dat}{a list of data consists of y_resid and d_reisd.}
}
\description{
CU-GMM variance function for RCS data
}
\keyword{internal}
