% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-main.R
\name{ddid_fit}
\alias{ddid_fit}
\title{Fit DID and sDID}
\usage{
ddid_fit(formula, data, lead = 1)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ Gi + It + Gi * It + x1 + x2}.}

\item{data}{An output of \code{did_panel_data()}.}
}
\value{
A numeric vector that containts the DID estimate and the sequential DID estimate.
}
\description{
Fit DID and sDID
}
\keyword{internal}
