% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-main.R
\name{ddid_sa_data}
\alias{ddid_sa_data}
\title{Double DID Data Format for the Staggered Adoption}
\usage{
ddid_sa_data(var_treat, id_unit, id_time, data)
}
\description{
Double DID Data Format for the Staggered Adoption
}
\keyword{internal}
