% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dekadt2018}
\alias{dekadt2018}
\title{Dataset from de Kadt and Larreguy (2018)}
\format{
A panel data in \code{tbl} and \code{data.frame} format.
}
\usage{
data(dekadt2018)
}
\description{
Dataset from de Kadt and Larreguy (2018)
}
\references{
De Kadt, Daniel, and Horacio A. Larreguy.
 "Agents of the regime? Traditional leaders and electoral politics in South Africa."
 \emph{The Journal of Politics} 80.2 (2018): 382-399.
 \doi{10.1086/694540}
}
\keyword{dataset}
