% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{did}
\alias{did}
\title{DiD}
\usage{
did(formula, data, id_subject, id_time, post_treatment,
  method = "parametric")
}
\arguments{
\item{formula}{A formula. \code{outcome ~ treatment}. Covariates can be added by \code{outcome ~ treatment | covariates}.}

\item{data}{\code{\link{did_data}} object}

\item{method}{Either \code{"parametric"} or \code{"nonparametric"}.}
}
\description{
DiD
}
