% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check.R
\name{did_check}
\alias{did_check}
\title{Assessing Assumptions for Difference-in-differences}
\usage{
did_check(
  formula,
  data,
  id_unit,
  id_time,
  design = "did",
  is_panel = TRUE,
  option = list()
)
}
\arguments{
\item{formula}{A formula of the following form,
\itemize{
  \item When \code{is_panel = TRUE}: \code{y ~ treatment | x1 + x2} where
   \code{treatment} is a time-varying treatment indicator.

  \item When \code{is_panel = FALSE}: \code{y ~ treat_group + post_treat | x1 + x2},
   where \code{treat_group} is a time-invariant treatment indicator,
   and \code{post_treat} is an indicator that takes 1 if the observations is from the
   post treatment periods and zero otherwise. The order is strict.

   \item Covariates are specified after the vertical bar \code{|}.
    When there are no covariates, the formula should be given as
    \code{y ~ treatment} for \code{is_panel = TRUE} and
    \code{y ~ treat_group + post_treat} for \code{is_panel = FALSE}.
}}

\item{data}{A data frame.}

\item{id_unit}{A variable name of unit (e.g., country name, respondent id).
When \code{is_panel = TRUE} (i.e., the panel data), a variable name of the unit index should be provided.
When \code{is_panel = FALSE} (i.e., the repeated cross-section data), this can be left as \code{NULL}.}

\item{id_time}{A variable name of time (e.g., year).}

\item{design}{The design to be used: either \code{"did"} (the standard difference-in-differences design) or \code{"sa"} (the staggered adoption design).
The default is \code{"did"}.}

\item{is_panel}{A boolean argument. This should be \code{TRUE} when the dataset is panel (i.e., the same units are repeatedly observed over time); This should be \code{FALSE} when the dataset is the repeated cross-section (RCS) where different sets of units are observed at each time point.}

\item{option}{A list of the following option parameters:
\describe{
  \item{n_boot}{The number of bootstrap iterations.}
  \item{se_boot}{A boolean argument. If \code{TRUE} bootstrap-based critical value is used to construct the confidence intervals.
                   Default is \code{FALSE}}
  \item{id_cluster}{A variable name of a cluster index used to conduct clustered bootstraps.
                    If left unspecified, the unit level bootstrap will be conducted.}
  \item{parallel}{Parallel computing for the block bootstrap. Use \code{future} package.}
  \item{lead}{Lead parameter. Default is 0, which estimate the instantaneous treatment effect.}
}}
}
\description{
Assessing Assumptions for Difference-in-differences
}
