% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-sad.R
\name{did_check_sad}
\alias{did_check_sad}
\title{Staggered Adoption Design}
\usage{
did_check_sad(formula, data, id_subject, id_time, option)
}
\description{
Staggered Adoption Design
}
\keyword{internal}
