% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-std.R
\name{did_check_std}
\alias{did_check_std}
\title{DID check function for the basic design}
\usage{
did_check_std(formula, data, id_unit, id_time, is_panel = TRUE, option)
}
\arguments{
\item{formula}{Formula.}

\item{data}{Panel data.}

\item{id_unit}{A variable name for unit index. Required for panel data.}

\item{id_time}{A variable name for time index.}

\item{is_panel}{A boolean argument. Should be \code{FALSE} for the repeated cross-sectional data.}

\item{option}{A list of option parameters.}
}
\value{
A list of placebo estimates and plots.
}
\description{
DID check function for the basic design
}
\keyword{internal}
