% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-data.R
\name{did_data_panelL}
\alias{did_data_panelL}
\title{Data processing function for panel data (long format)}
\usage{
did_data_panelL(outcome, treatment, post_treatment, id_subject, id_time,
  Xcov, x_formula = NULL)
}
\arguments{
\item{outcome}{A vector of outcome variable.}

\item{treatment}{A vector of treatment indicator variable.}

\item{post_treatment}{A vector of numeric time index for post treatment periods.}

\item{id_subject}{A vector of subject index variable.}

\item{id_time}{A vector of time index variable.}

\item{Xcov}{A matrix of covariates.}
}
\value{
A list. Each element corresponds to each post-treatment period.
 Each element of the returned list is also a list consists of the following:
\itemize{
 \item \code{Y}: outcome vector.
 \item \code{D}: treatment vector.
 \item \code{formula}: A list of formula for \code{\link{lm}}.
 \item \code{pdata}: A data frame appended transformed outcomes (e.g., yd0, yd1, etc).
}
}
\description{
Data processing function for panel data (long format)
}
\keyword{internal}
