% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-data.R
\name{did_data_panelW}
\alias{did_data_panelW}
\title{Data processing function for panel data (wide format)}
\usage{
did_data_panelW(outcome, treatment, post_treat)
}
\arguments{
\item{outcome}{A vector of outcome variable.}

\item{treatment}{A vector of treatment indicator variable.}

\item{post_treatment}{A vector of numeric time index for post treatment periods.}
}
\value{
A list. Each element corresponds to each post-treatment period.
 Each element of the returned list is also a list consists of the following:
\itemize{
 \item \code{Y}: outcome vector.
 \item \code{D}: treatment vector.
}
}
\description{
Data processing function for panel data (wide format)
}
\keyword{internal}
