% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-data_old.R
\name{did_data_rcs}
\alias{did_data_rcs}
\title{Data processing function for repeated cross-section data}
\usage{
did_data_rcs(outcome, treatment, post_treatment, id_time, Xcov,
  x_formula = NULL)
}
\arguments{
\item{outcome}{a vector of outcome variable.}

\item{treatment}{a vector of treatment indicator variable.}

\item{post_treatment}{a vector of numeric time index for post treatment periods.}

\item{id_time}{a vector of time index variable.}

\item{Xcov}{a matrix of covariates.}
}
\value{
\code{\link{did_data_rcs}} returns a list. Each element corresponds to each post-treatment period.
 Each element of the returned list is also a list consists of the following:
\itemize{
 \item \code{Y}: outcome vector.
 \item \code{D}: treatment vector.
 \item \code{formula}: A list of formula for \code{\link{lm}}.
 \item \code{pdata}: A data frame appended transformed outcomes (e.g., yd0, yd1, etc).
}
}
\description{
Data processing function for repeated cross-section data
}
\keyword{internal}
