% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-functions.R
\name{did_formula}
\alias{did_formula}
\title{Transform formula}
\usage{
did_formula(formula, is_panel)
}
\arguments{
\item{formula}{Formula.}

\item{is_panel}{A boolean argument to indicate if data is in the panel format.}
}
\value{
A list of working objects.
}
\description{
Transform formula
}
\keyword{internal}
