% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-nonparametric.R
\name{did_nonparametric}
\alias{did_nonparametric}
\title{Double Difference-in-Differences Estimator}
\usage{
did_nonparametric(data, se_boot = FALSE, n_boot = 1000,
  boot_min = TRUE, select = "HQIC", est_did = TRUE)
}
\arguments{
\item{data}{did_double.data object. See \code{\link{did_double_data}}.}

\item{se_boot}{A boolean argument.
If set to \code{TRUE}, standard errors are computed by the block bootstrap.
Bootstrap iterations are set by \code{n_boot} argument.}

\item{n_boot}{The number of bootstrap iterations. Required when \code{se_boot == TRUE}.}

\item{boot_min}{If \code{TRUE}, bootstrap is used only for the selected model.
This option helps reduce computational burdens.}

\item{select}{The criteria used to select the best model. The selected model is used to estimate bootstrap variance when \code{boot_min = TRUE}.
Options are "HQIC", "BIC", "tt1" (T-test) and "tt2" (T-test with Bonferroni correction).}

\item{est_did}{If \code{TRUE}, standard difference-in-differences estimates are computed. Default is \code{TRUE}.}
}
\description{
Double Difference-in-Differences Estimator
}
\examples{
 # load packages
 require(didrobust)

 # load data from didrobust
 data(anzia2012)

 # transform data
 dat1 <- did_double_data(
   outcome = anzia2012$lnavgsalary_cpi,
   treatment = anzia2012$oncycle,
   post_treatment = c(2007, 2008, 2009),
   id_subject = anzia2012$district,
   id_time = anzia2012$year
 )

 # fit the model
 set.seed(1234)
 fit <- did_double(dat1, se_boot = TRUE, n_boot = 100, select = "HQIC")
 summary(fit)
 plot(fit)

 # run bootstrap for all models
 set.seed(1234)
 fit2 <- did_double(dat1, se_boot = TRUE, n_boot = 100, boot_min = FALSE)
 plot(fit2, ylim = c(-0.05, 0.05), full = TRUE)
 abline(h = 0, col = 'red', lty = 3)
}
