% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric.R
\name{did_parametric}
\alias{did_parametric}
\title{Parametric Version of Double DiD}
\usage{
did_parametric(data, se_boot = FALSE, n_boot = 1000, boot_min = TRUE,
  select = "HQIC", est_did = TRUE)
}
\arguments{
\item{data}{data}

\item{se_boot}{A boolean argument.
If set to \code{TRUE}, standard errors are computed by the block bootstrap.
Bootstrap iterations are set by \code{n_boot} argument.}

\item{n_boot}{The number of bootstrap iterations. Required when \code{se_boot == TRUE}.}

\item{boot_min}{If \code{TRUE}, bootstrap is used only for the selected model.
This option helps reduce computational burdens.}

\item{select}{The criteria used to select the best model. The selected model is used to estimate bootstrap variance when \code{boot_min = TRUE}.
Options are "HQIC", "BIC", "tt1" (T-test) and "tt2" (T-test with Bonferroni correction).}
}
\description{
Parametric Version of Double DiD
}
