% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{did_parametric_rcs}
\alias{did_parametric_rcs}
\title{Double DiD with Repeated Cross-Section}
\usage{
did_parametric_rcs(data)
}
\arguments{
\item{data}{A \code{diddesign_data} object.}
}
\description{
Double DiD with Repeated Cross-Section
}
