% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{did_parametric_repeatedCS}
\alias{did_parametric_repeatedCS}
\title{Double DiD with Repeated Cross-Section}
\usage{
did_parametric_repeatedCS(data, x_colnames)
}
\arguments{
\item{data}{A \code{diddesign_data} object.}

\item{x_colnames}{A vector of covariate variable names. 
Parsed from formula in \code{\link{did}} function.}
}
\description{
Double DiD with Repeated Cross-Section
}
