% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{did_plot}
\alias{did_plot}
\title{DiD plot}
\usage{
did_plot(formula, data, post_treatment, id_subject = NULL, id_time,
  diff_order = 1, xlim = NULL, ylim = NULL, col = NULL,
  loc = NULL, lwd = NULL, ...)
}
\description{
DiD plot
}
\examples{
# load package
require(DIDdesign)

# load data
data(anzia2012)

# plot
did_plot(lnavgsalary_cpi ~ oncycle, data = anzia2012,
  post_treatment = c(2007, 2008, 2009),
  id_subject = "district",
  id_time = "year",
  ylim = c(10.5, 10.8)
)
}
