% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{did_plot_selection}
\alias{did_plot_selection}
\title{Plot function for selection}
\usage{
did_plot_selection(data, alpha = 0.05, equivalence = TRUE, eL = NULL,
  eU = NULL, xlim = NULL, ylim = NULL, col = NULL, loc = NULL,
  ...)
}
\arguments{
\item{data}{\code{diddesign} object.
Typically, it is an output from \code{\link{did}} function.}

\item{alpha}{Level of rejection.
Confidence intervals are constructed with the 1-alpha and 1 - alpha/2 levels.}
}
\description{
Plot function for selection
}
\examples{
# load package
require(DIDdesign)

# load data
data(anzia2012)

# fit the model
fit1 <- did(lnavgsalary_cpi ~ oncycle, data = anzia2012,
            id_subject = "district", id_time = "year",
            post_treatment = c(2007, 2008, 2009),
            method = "parametric")

# plot
did_plot_selection(fit1, equivalence = TRUE, eL = -0.005, eU = 0.005)
}
\keyword{internal}
