% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-data_prep.R
\name{did_rcs_data}
\alias{did_rcs_data}
\title{Prepare data for the DID design with repeated cross-section data}
\usage{
did_rcs_data(
  var_outcome,
  var_treat,
  var_post,
  var_covars = NULL,
  var_cluster = NULL,
  id_time,
  data
)
}
\description{
Prepare data for the DID design with repeated cross-section data
}
\keyword{internal}
