% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{did_sad}
\alias{did_sad}
\title{Function to implement DDID under the Staggered Adoption Design}
\usage{
did_sad(formula, data, id_subject, id_time, option)
}
\arguments{
\item{formula}{A formula indicating outcome and treatment.}

\item{data}{A long form panel data.}

\item{id_subject}{A character variable indicating subject index.}

\item{id_time}{A character variable indicating time index.}
}
\description{
Function to implement DDID under the Staggered Adoption Design
}
\keyword{internal}
