% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-sad.R
\name{did_sad_pattern}
\alias{did_sad_pattern}
\title{Generate a pattern plot}
\usage{
did_sad_pattern(data, treatment, Gmat)
}
\arguments{
\item{data}{Panel data.}

\item{treatment}{Name of the treatment variable.}

\item{Gmat}{A matrix of treatment patterns.}
}
\value{
A list of the treatment variation plot and the data used to generate the plot.
}
\description{
Generate a pattern plot
}
\keyword{internal}
