% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-sad.R
\name{did_sad_placebo}
\alias{did_sad_placebo}
\title{SA Placebo Regression}
\usage{
did_sad_placebo(fm_prep, dat_panel, treatment, outcome, option)
}
\description{
SA Placebo Regression
}
\keyword{internal}
