% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-sad.R
\name{did_sad_plot}
\alias{did_sad_plot}
\title{Create a did plot for standard design}
\usage{
did_sad_plot(data)
}
\arguments{
\item{data}{A data frame of double did estimates.}
}
\value{
A ggplot2 object
}
\description{
Create a did plot for standard design
}
\keyword{internal}
