% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-std.R
\name{did_std}
\alias{did_std}
\title{Implement the Double DID under the standard design}
\usage{
did_std(formula, data, id_unit, id_time, is_panel = TRUE, option)
}
\description{
Implement the Double DID under the standard design
}
\keyword{internal}
