% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-std.R
\name{did_std_placebo}
\alias{did_std_placebo}
\title{Run a placebo regression on the pretreatment outcome}
\usage{
did_std_placebo(formula, data, lags, skip_standardize, show_warning = TRUE)
}
\arguments{
\item{formula}{A formula generated by \code{did_formula} function.}

\item{data}{An output from \code{did_panel_data} or \code{did_rcs_data} function.}

\item{lags}{A vector of non-negative lag parameters.}

\item{skip_standardize}{A logical value indicating whether to skip standardization.}

\item{show_warning}{A logical value indicating whether to show warnings.}
}
\value{
A list of placebo effects (\code{est}),
  and standardized effects (\code{est_std})
}
\description{
Run a placebo regression on the pretreatment outcome
}
\keyword{internal}
