% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check.R
\name{did_std_plot}
\alias{did_std_plot}
\title{Create a did plot for standard design}
\usage{
did_std_plot(data)
}
\arguments{
\item{data}{A data object from \code{did_panel_data} or \code{did_rcs_data}}
}
\value{
A ggplot object
}
\description{
Create a did plot for standard design
}
\keyword{internal}
