% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm_trend_test.R
\name{didgmm_test2}
\alias{didgmm_test2}
\title{Pre-treatment Trend test with GMM}
\usage{
didgmm_test2(Y0, D0, M, max_trial = 100)
}
\description{
Pre-treatment Trend test with GMM
}
\keyword{internal}
