% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmmT.R
\name{didgmmT}
\alias{didgmmT}
\title{DID-GMM with m-th order parallel treand assumption}
\usage{
didgmmT(
  Y,
  D,
  M,
  ep = 0.01,
  max_trial = 100,
  only_beta = FALSE,
  only_oneM = FALSE
)
}
\arguments{
\item{Y}{Outcome matrix.}

\item{D}{Treatment vector.}

\item{M}{Order specification.}

\item{ep}{small constant added to diagonal elements of the weighting matrix in GMM. Default is 0.01.}

\item{only_beta}{A boolean argument.
If \code{TRUE} ATT is estimated only using moment conditions related to the last time-period.
Default is \code{FALSE}.}

\item{only_oneM}{A boolean argument.
if \code{TRUE}, ATT is estimated only using one moment condition.}
}
\value{
A list consisting of the following elements:
 \itemize{
   \item \code{ATT}: Estimates of the average treatment effect on the treated.
   \item \code{zeta}: a value of zeta.
   \item \code{J}: J-statistics
   \item \code{BIC}: BIC
   \item \code{HQIC}: HQIC
 }
}
\description{
A function for DID-GMM with panel data.
}
\keyword{internal}
