% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric.R
\name{didgmmT_parametric}
\alias{didgmmT_parametric}
\title{Estimate parameter by CU-GMM}
\usage{
didgmmT_parametric(dat, id_subject, par_init = NULL)
}
\description{
Estimate parameter by CU-GMM
}
\keyword{internal}
