% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{didgmmT_parametric_rcs}
\alias{didgmmT_parametric_rcs}
\title{CU-GMM function for RCS data}
\usage{
didgmmT_parametric_rcs(dat, par_init = NULL, optim = TRUE,
  gmm = "CU")
}
\arguments{
\item{A}{list of dataset. Output of \code{link{getX_rcs}}.}

\item{Xdesign}{a list of design matrix. The variable of interest is named 'treatment:post'}
}
\description{
CU-GMM function for RCS data
}
\keyword{internal}
