% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric.R
\name{didgmmT_parametric_resid}
\alias{didgmmT_parametric_resid}
\title{Estimate GMM with residualized outcome / treatment}
\usage{
didgmmT_parametric_resid(dat, id_subject, par_init = NULL)
}
\description{
Estimate GMM with residualized outcome / treatment
}
\keyword{internal}
