% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-main.R
\name{double_did_compute}
\alias{double_did_compute}
\title{Compute Point and Variance Estimates}
\usage{
double_did_compute(fit, weights, se_boot = TRUE)
}
\description{
Compute Point and Variance Estimates
}
\keyword{internal}
