% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm-selection.R
\name{fe_selection}
\alias{fe_selection}
\title{Model selectin based on linear FE model}
\usage{
fe_selection(data, fm_list, post_time, alpha = 0.05)
}
\arguments{
\item{data}{panel data from \code{diddesign_data} object.}

\item{fm_list}{A list of formulae from \code{diddesign_data} object.}

\item{post_time}{Time index for the post Treatment period.}

\item{alpha}{Level of rejection. Default is at the 5\% level}
}
\description{
Model selectin based on linear FE model
}
\keyword{internal}
