% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{gen_dummies}
\alias{gen_dummies}
\title{Generate dummy variables}
\usage{
gen_dummies(data, add_intercept = TRUE)
}
\arguments{
\item{data}{An element of \code{diddesign_data} object.}
}
\value{
A data matrix. The first column is ones' if \code{add_intercept = TRUE}.
}
\description{
Generate dummy variables
}
\keyword{internal}
