% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{generate_tab_parametric}
\alias{generate_tab_parametric}
\title{Generate Table for Parametric Method}
\usage{
generate_tab_parametric(obj, ATT, se_save, selected, full = TRUE)
}
\description{
Generate Table for Parametric Method
}
\keyword{internal}
