% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{get_periods}
\alias{get_periods}
\title{Obtain periods used for the analysis}
\usage{
get_periods(Gmat, thres = 2)
}
\arguments{
\item{Gmat}{G matrix produced in \code{create_Gmat()}.}

\item{thres}{A minimum number of treatd units for the period included in the analysis. Default is 2.}
}
\description{
Obtain periods used for the analysis
}
\keyword{internal}
