% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad-functions.R
\name{get_subjects}
\alias{get_subjects}
\title{Obtain subject index for each periods}
\usage{
get_subjects(Gmat, id_time_use)
}
\arguments{
\item{id_time_use}{A vector of time index. Should be normalized.}

\item{Gamt}{G matrix created by \code{create_Gmat()}.}
}
\description{
Obtain subject index for each periods
}
\keyword{internal}
