% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad-functions.R
\name{get_time_weight}
\alias{get_time_weight}
\title{Compute time-sepcific weights}
\usage{
get_time_weight(Gmat, id_time_use)
}
\arguments{
\item{Gmat}{G matrix.}

\item{id_time_use}{A vector of time index (normalized). Output of \code{get_periods}.}
}
\description{
Compute time-sepcific weights
}
\keyword{internal}
