% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{getX_rcs}
\alias{getX_rcs}
\title{get design matrix and outcome vector}
\usage{
getX_rcs(lm_fit, response_orginal)
}
\arguments{
\item{lm_fit}{Fitted output from \code{\link{lm}}.}
}
\description{
get design matrix and outcome vector
}
\keyword{internal}
