% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm-selection.R
\name{gmm_selection}
\alias{gmm_selection}
\title{Model selection by GMM}
\usage{
gmm_selection(Y, D, mvec, t_pre, select, n_boot)
}
\arguments{
\item{Y}{outcome}

\item{D}{treatment}

\item{select}{selection methods}

\item{m_vec}{moment sets}
}
\description{
Model selection by GMM
}
\keyword{internal}
