% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmmT.R
\name{hansenT_over_onlyB_ver2}
\alias{hansenT_over_onlyB_ver2}
\title{Main loss function for Case 3 (only beta)}
\usage{
hansenT_over_onlyB_ver2(par, Y, D, M, init = NULL)
}
\description{
Main loss function for Case 3 (only beta)
}
