% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm_trend_test.R
\name{hansenT_tretest_over}
\alias{hansenT_tretest_over}
\title{compute loss based on moment condition}
\usage{
hansenT_tretest_over(par, Y, D, M, ep)
}
\description{
compute loss based on moment condition
}
\keyword{internal}
