% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm_trend_test.R
\name{hansenT_tretest_over2}
\alias{hansenT_tretest_over2}
\title{compute loss based on moment condition}
\usage{
hansenT_tretest_over2(par, Y0, D0, M, init = FALSE, return_var = FALSE)
}
\description{
compute loss based on moment condition
}
\keyword{internal}
