% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm_trend_test.R
\name{hansenT_tretest_over_init}
\alias{hansenT_tretest_over_init}
\title{function to select intial value}
\usage{
hansenT_tretest_over_init(par, Y, D, M, ep)
}
\description{
function to select intial value
}
\keyword{internal}
