% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm_trend_test.R
\name{hansenT_tretest_variance}
\alias{hansenT_tretest_variance}
\title{GMM variance for the testing moment model}
\usage{
hansenT_tretest_variance(par, Y, D, M)
}
\description{
GMM variance for the testing moment model
}
\keyword{internal}
