% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_loop}
\alias{loss_loop}
\title{helper function for the loss computation}
\usage{
loss_loop(X, y, par, id_subject, uid, is_na, nobs, p)
}
\description{
helper function for the loss computation
}
\keyword{internal}
