% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paglayan2019}
\alias{paglayan2019}
\title{Dataset from Paglayan (2019)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2142 rows and 5 columns.
}
\usage{
paglayan2019
}
\description{
Dataset from Paglayan (2019)
}
\keyword{dataset}
