% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{plot.DIDdesign}
\alias{plot.DIDdesign}
\title{Plot output from did function}
\usage{
\method{plot}{DIDdesign}(x, check_fit = NULL, band = FALSE, ...)
}
\arguments{
\item{x}{An output from \code{did} function.}

\item{check_fit}{An output from \code{did_check} function.}

\item{band}{A boolean argument. If \code{TRUE}, confidence intervals are shown with \code{geom_ribbon} function.}

\item{...}{Other parameters. Currently not supported.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot output from did function
}
