% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check.R
\name{plot.DIDdesign_check}
\alias{plot.DIDdesign_check}
\title{Plot output from did_check}
\usage{
\method{plot}{DIDdesign_check}(x, ...)
}
\arguments{
\item{x}{An output from \code{did_check} function.}

\item{...}{Other parameters. Currently not supported.}
}
\description{
Plot output from did_check
}
