% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{plot_diddesign_data}
\alias{plot_diddesign_data}
\title{Baseline plot function for did_plot}
\usage{
plot_diddesign_data(data, panel = TRUE, xlim = NULL, ylim = NULL,
  col = NULL, loc = NULL, lwd = NULL, ...)
}
\arguments{
\item{data}{Two possible inputs. If \code{panel = TRUE}, input should be a \code{diddesign_data} class object.
If \code{panel = FALSE}, it should be a list consists of \code{"y1mean"}, \code{"y0mean"} and \code{"id_time"}.}
}
\description{
Baseline plot function for did_plot
}
\keyword{internal}
