% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{print.summary.DIDdesign_check}
\alias{print.summary.DIDdesign_check}
\title{Print}
\usage{
\method{print}{summary.DIDdesign_check}(x, ...)
}
\arguments{
\item{x}{An ouput of \code{summary.DIDdesign_check} function.}

\item{...}{Other parameters. Currently not supported.}
}
\description{
Print
}
