% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{print.summary.diddesign_data}
\alias{print.summary.diddesign_data}
\title{Print output from summary function}
\usage{
\method{print}{summary.diddesign_data}(obj)
}
\arguments{
\item{obj}{an object of \code{summary.diddesign_data} class, typically an ouput from \code{\link{summary.diddesign}}.}
}
\description{
Print output from summary function
}
