% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcs_mean_fill}
\alias{rcs_mean_fill}
\title{helper function for repeated cross-section data}
\usage{
rcs_mean_fill(outcome, treatment, time_index, time_unique, y1mean, y0mean)
}
\arguments{
\item{outcome}{outcome vector}

\item{treatment}{treatment vector}

\item{time_index}{time id for each unit}

\item{time_unique}{a vector of unique time index (should be increasing order)}

\item{y1mean}{a vector of mean outcome for the treated}

\item{y0mean}{a vector of mean outcome for the control}
}
\value{
a matrix of outcome filled with mean outcome for each group
}
\description{
helper function for repeated cross-section data
}
\keyword{internal}
