% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{sa_calc_cov}
\alias{sa_calc_cov}
\title{Compute lead specific variance covariance matrix}
\usage{
sa_calc_cov(obj, lead)
}
\description{
Compute lead specific variance covariance matrix
}
\keyword{internal}
