% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{sa_calc_ddid_var}
\alias{sa_calc_ddid_var}
\title{Compute Bootstrap-based Variance for the SA-Double-DID Estimator}
\usage{
sa_calc_ddid_var(obj, lead, w_did, w_sdid)
}
\arguments{
\item{obj}{A list of bootstrap outputs.}

\item{lead}{A vector of the lead parameter.}

\item{w_did}{Estimated weight for the DID estimator}

\item{w_sdid}{Estimated weight for the sequential DID estimator}
}
\description{
Compute Bootstrap-based Variance for the SA-Double-DID Estimator
}
\keyword{internal}
