% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{sa_did}
\alias{sa_did}
\title{Function to implement the Staggered Adption Design}
\usage{
sa_did(formula, data, id_subject = NULL, id_time)
}
\description{
Function to implement the Staggered Adption Design
}
