% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{sa_did_est}
\alias{sa_did_est}
\title{Function to implement the Staggered Adption Design}
\usage{
sa_did_est(formula, data, id_subject, id_time)
}
\arguments{
\item{formula}{A formula indicating outcome and treatment.}

\item{data}{A long form panel data.}

\item{id_subject}{A character variable indicating subject index.}

\item{id_time}{A character variable indicating time index.}
}
\description{
Function to implement the Staggered Adption Design
}
\examples{
# simulate data 
set.seed(1234)
dat <- simulate_sad(30, 10)

# estimate 
sa_did_est(outcome ~ treatment, data = dat$dat, 
  id_subject = "id_subject", id_time = 'id_time'
)
}
