% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{sa_did_to_ddid}
\alias{sa_did_to_ddid}
\title{Convert estimates into double did}
\usage{
sa_did_to_ddid(obj_point, obj_boot, lead)
}
\arguments{
\item{obj_point}{Point estimates.}

\item{obj_boot}{Bootstrap outputs.}

\item{lead}{A vector of the lead parameter.}
}
\value{
A list of estimates and weights.
}
\description{
Convert estimates into double did
}
\keyword{internal}
