% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{sample_panel}
\alias{sample_panel}
\title{Block Bootstrap used in SA design}
\usage{
sample_panel(panel_dat)
}
\arguments{
\item{panel_dat}{A panel data.}
}
\value{
A panel data, sampled from the input data.
}
\description{
Block Bootstrap used in SA design
}
\keyword{internal}
