% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sances2015}
\alias{sances2015}
\title{Dataset from Sances (2015)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23000 rows and 6 columns.
}
\usage{
sances2015
}
\description{
Dataset from Sances (2015)
}
\references{
Michael W. Sances.
"The Distributional Impact of Greater Responsiveness."
 \emph{The Journal of Politics} 78.1 (2016).
 \doi{10.1086/683026}
}
\keyword{dataset}
