% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-functions.R
\name{set_option}
\alias{set_option}
\title{Set default value of options}
\usage{
set_option(option)
}
\arguments{
\item{option}{A list of options.}
}
\value{
A list of options.
}
\description{
Set default value of options
}
\keyword{internal}
