% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-functions.R
\name{setup_parallel}
\alias{setup_parallel}
\title{Setup Parallel Workers}
\usage{
setup_parallel(is_parallel)
}
\arguments{
\item{is_parallel}{A boolean argument. If \code{TRUE} bootstrap will be conducted on multiple cores.}
}
\description{
Setup Parallel Workers
}
\keyword{internal}
