% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-selection.R
\name{sign_test}
\alias{sign_test}
\title{Function to test directin of trends.}
\usage{
sign_test(dat, level = 0.1)
}
\arguments{
\item{dat}{A data object of a class \code{diddesign_data}. Typically an output from \code{did_data}.}

\item{level}{level of rejection. Default is 0.1.}
}
\description{
This function test if pre-treatment trends are in the same direction. 
Currently we support when there are only two pre-treatment time points.
}
\examples{
# load data from didrobust package
data(anzia2012)

# create did_data object 
out <- did_data(
  outcome = anzia2012$lnavgsalary_cpi,
  treatment = anzia2012$oncycle,
  post_treatment = c(2007, 2008, 2009),
  id_subject = anzia2012$district,
  id_time = anzia2012$year,
  long = TRUE
)

# run the test 
tt <- sign_test(out)
tt$pvalue 
}
