% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-selection.R
\name{sign_test_panel}
\alias{sign_test_panel}
\title{Function to test directin of trends for panel}
\usage{
sign_test_panel(dat, level = 0.1)
}
\description{
Function to test directin of trends for panel
}
\keyword{internal}
