% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-selection.R
\name{sign_test_rcs}
\alias{sign_test_rcs}
\title{Function to test direction of trends for RCS}
\usage{
sign_test_rcs(dat, level = 0.1)
}
\description{
Function to test direction of trends for RCS
}
\keyword{internal}
