% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad-simulate.R
\name{simulate_sad}
\alias{simulate_sad}
\title{Generate staggered adoption design}
\usage{
simulate_sad(n_obs, n_times)
}
\arguments{
\item{n_obs}{Number of units.}

\item{n_time}{Number of time periods.}
}
\value{
\code{simulate_sad()} returns a list containing the following elements:
\itemize{
  \item{dat} A long format dataset generated under the staggered adoption.
  \item{Amat} A matrix of treatment patterns. 
}
}
\description{
Generate staggered adoption design
}
\keyword{internal}
