% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{summary.DIDdesign}
\alias{summary.DIDdesign}
\title{Summarize DIDdesign output}
\usage{
\method{summary}{DIDdesign}(object, estimator = NULL, ...)
}
\arguments{
\item{object}{An object of \code{DIDdesign} class.}

\item{estimator}{A vector of estimators to print. For example, \code{estimator = c("DID", "sDID")}.}

\item{...}{Other parameters. Currently not supported.}
}
\description{
Summarize DIDdesign output
}
