% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{summary.DIDdesign_check}
\alias{summary.DIDdesign_check}
\title{Summary did_check Output}
\usage{
\method{summary}{DIDdesign_check}(object, ...)
}
\arguments{
\item{object}{An output of \code{did_check} function.}

\item{...}{Other parameters. Currently not supported.}
}
\description{
Summary did_check Output
}
