% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{var_boot_parametric_rcs}
\alias{var_boot_parametric_rcs}
\title{Varinace for Repated Cross-Section with Block-Bootstrap}
\usage{
var_boot_parametric_rcs(dat, n_boot, id_cluster, verbose)
}
\arguments{
\item{n_boot}{number of bootstrap iterations}

\item{id_cluster}{cluster index at the level of blocking}

\item{data}{a list of data}
}
\description{
Varinace for Repated Cross-Section with Block-Bootstrap
}
\keyword{internal}
