% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anzia2012}
\alias{anzia2012}
\title{Dataset from Anzia (2012)}
\format{A \code{tbl} format data.}
\usage{
data(anzia2012)
}
\description{
Dataset from Anzia (2012)
}
\references{
Anzia, Sarah F.
 "The Election Timing Effect: Evidence from a Policy Intervention in Texas."
 \emph{Quarterly Journal of Political Science} 7.3 (2012):209-248.
 \doi{10.1561/100.00011056}
}
\keyword{dataset}
