% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{compute_did}
\alias{compute_did}
\title{Estimate DID}
\usage{
compute_did(
  dat_panel,
  outcome,
  treatment,
  id_time_use,
  id_subj_use,
  time_weight,
  min_time = 3,
  lead
)
}
\description{
Estimate DID
}
\keyword{internal}
