% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{did}
\alias{did}
\title{DiD}
\usage{
did(data, method = "parametric")
}
\arguments{
\item{data}{\code{\link{did_data}} object}

\item{method}{Either \code{"parametric"} or \code{"nonparametric"}.}
}
\description{
DiD
}
