% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{did}
\alias{did}
\title{Double Difference-in-Differences}
\usage{
did(formula, data, id_subject = NULL, id_time, post_treatment,
  method = "parametric", se_boot = FALSE, n_boot = 1000,
  boot_min = TRUE, select = "GMM")
}
\arguments{
\item{formula}{A formula. \code{y ~ d} where \code{y} is the outcome and \code{d} is the (time-varying) treatment indicator.
Covariates can be added by \code{y ~ d | x1 + x2} where \code{x1} and \code{x2} are covariates.}

\item{data}{Data matrix. Either a \code{data.frame} object or  \code{tbl_df} object.}

\item{id_subject}{subject id. This argument should be left as \code{NULL} when repeated cross-section data are used.}

\item{id_time}{time id.}

\item{post_treatment}{time index for post treatment periods.}

\item{method}{Either \code{"parametric"} or \code{"nonparametric"}.}

\item{se_boot}{If \code{TRUE} blockbootstrap is used for computing standard errors.}

\item{n_boot}{Number of bootstrap iterations.}

\item{boot_min}{If \code{TRUE} bootstrap is carried out only for the selected model.}

\item{select}{Selection criteria.}
}
\value{
A \code{diddesign} class object, which is a list of results for each post treatment period.
 A result for one period contains:
 \itemize{
   \item \code{results_estimates}:
   \item \code{results_bootstraps}: bootstrap estimates.
   \item \code{BIC}: BIC for all models .
   \item \code{HQIC}: HQIC for all models.
   \item \code{BIC_min}: BIC for the selected model.
     This is the smallest BIC among values reported in \code{BIC}.
   \item \code{HQIC_min}: HQIC for the selected model.
     This is the smallest HQIC among values reported in \code{HQIC}.
   \item \code{min_model}: Selected model.
   \item \code{select}: A criteria used to select the model.
   \item \code{ATT}: Estimated average treatment effect on the treated.
   \item \code{ci95}: 95\% confidence intervals.
   \item \code{ci90}: 90\% confidence intervals.
 }
}
\description{
Estimate average treatment effect on the treated using double difference-in-difference estimator.
}
\examples{
  # load package
  require(DIDdesign)

  # load  data
  data(anzia2012)

  # nonparametric estimator without covariates
  fit1 <- did(
    formula = lnavgsalary_cpi ~ oncycle,
    data = anzia2012,
    id_subject = "district", id_time = "year",
    post_treatment = c(2007, 2008, 2009),
    method = "nonparametric",
    se_boot = TRUE, n_boot = 100, boot_min = FALSE
  )

  # view summary
  summary(fit1)

  # view effect plot
  plot(fit1, full = TRUE)
}
