% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-sad.R
\name{did_check_sad}
\alias{did_check_sad}
\title{Staggered Adoption Design}
\usage{
did_check_sad(formula, data, id_subject, id_time, option)
}
\arguments{
\item{formula}{A formula indicating outcome and treatment.}

\item{data}{A long form panel data.}

\item{id_subject}{A character variable indicating subject index.}

\item{id_time}{A character variable indicating time index.}
}
\value{
A list of placebo estimates and plots.
}
\description{
Staggered Adoption Design
}
\keyword{internal}
