% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-std.R
\name{did_check_std}
\alias{did_check_std}
\title{DID check function for the basic design}
\usage{
did_check_std(formula, data, id_unit, id_time, is_panel = TRUE, option)
}
\description{
DID check function for the basic design
}
\keyword{internal}
