% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-std.R
\name{did_compute_weights}
\alias{did_compute_weights}
\title{Compute Weighting Matrix via Bootstrap}
\usage{
did_compute_weights(fm_prep, dat_did, var_cluster, is_panel, option)
}
\description{
Compute Weighting Matrix via Bootstrap
}
\keyword{internal}
