% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_data.R
\name{did_data_rcs}
\alias{did_data_rcs}
\title{data processing function for repeated cross-section data}
\usage{
did_data_rcs(outcome, treatment, post_treatment, id_time, Xcov)
}
\description{
data processing function for repeated cross-section data
}
\keyword{internal}
