% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-nonparametric.R
\name{did_nonparametric}
\alias{did_nonparametric}
\title{Nonparametric Double Difference-in-Differences Estimator}
\usage{
did_nonparametric(data, se_boot = FALSE, n_boot = 1000,
  boot_min = TRUE, alpha = 0.05, select = "GMM", est_did = TRUE,
  verbose = TRUE)
}
\arguments{
\item{data}{did_double.data object. See \code{\link{did_double_data}}.}

\item{se_boot}{A boolean argument.
If set to \code{TRUE}, standard errors are computed by the block bootstrap.
Bootstrap iterations are set by \code{n_boot} argument.}

\item{n_boot}{The number of bootstrap iterations. Required when \code{se_boot == TRUE}.}

\item{boot_min}{If \code{TRUE}, bootstrap is used only for the selected model.
This option helps reduce computational burdens.}

\item{select}{The criteria used to select the best model. The selected model is used to estimate bootstrap variance when \code{boot_min = TRUE}.
Options are "GMM", "tt1" (T-test) and "tt2" (T-test with Bonferroni correction).}

\item{est_did}{If \code{TRUE}, standard difference-in-differences estimates are computed. Default is \code{TRUE}.}
}
\description{
Nonparametric Double Difference-in-Differences Estimator
}
\examples{
\donttest{
 # load package
 require(DIDdesign)
 # load data
 data(anzia2012)

 # convert data into did double format
 out <- did_data(
   outcome = anzia2012$lnavgsalary_cpi,
   treatment = anzia2012$oncycle,
   post_treatment = c(2007, 2008, 2009),
   id_subject = anzia2012$district,
   id_time = anzia2012$year,
   long = TRUE, Xcov = NULL
 )

 # fit nonparametric double-did method
 fit <- did_nonparametric(out)

 # plot results
 par(mar = c(4, 2.5, 3.5, 1))
 plot(fit, full = TRUE, ylim = c(-0.05, 0.05))

 # show result summary
 summary(fit)
}
}
\seealso{
Other estimation functions: \code{\link{did_parametric_rcs}},
  \code{\link{did_parametric}}
}
\concept{estimation functions}
