% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric.R
\name{did_parametric}
\alias{did_parametric}
\title{Parametric Double Difference-in-Differences Estimator}
\usage{
did_parametric(data, se_boot = FALSE, n_boot = 1000, boot_min = TRUE,
  select = "parametric", est_did = FALSE, is_covariates,
  verbose = TRUE)
}
\arguments{
\item{data}{data}

\item{se_boot}{A boolean argument.
If set to \code{TRUE}, standard errors are computed by the block bootstrap.
Bootstrap iterations are set by \code{n_boot} argument.}

\item{n_boot}{The number of bootstrap iterations. Required when \code{se_boot == TRUE}.}

\item{boot_min}{If \code{TRUE}, bootstrap is used only for the selected model.
This option helps reduce computational burdens.
The selected model is used to estimate bootstrap variance when \code{boot_min = TRUE}.}

\item{selection}{either \code{"parametric"} or \code{'GMM'}.}
}
\description{
Parametric Double Difference-in-Differences Estimator
}
\seealso{
Other estimation functions: \code{\link{did_nonparametric}},
  \code{\link{did_parametric_rcs}}
}
\concept{estimation functions}
