% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-parametric-rcs.R
\name{did_parametric_rcs}
\alias{did_parametric_rcs}
\title{Parametric Double Difference-in-Differences Estimator with Repeated Cross-Section Data}
\usage{
did_parametric_rcs(data, verbose = TRUE)
}
\arguments{
\item{data}{A \code{diddesign_data} object.}
}
\description{
Parametric Double Difference-in-Differences Estimator with Repeated Cross-Section Data
}
\seealso{
Other estimation functions: \code{\link{did_nonparametric}},
  \code{\link{did_parametric}}
}
\concept{estimation functions}
