% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{did_plot}
\alias{did_plot}
\title{DiD plot}
\usage{
did_plot(outcome, treatment, post_treatment, id_subject, id_time,
  xlim = NULL, ylim = NULL, col = NULL, ...)
}
\description{
DiD plot
}
\examples{
# load package 
require(DIDdesign)

# load data 
data(anzia2012)

# plot
did_plot(
  outcome = anzia2012$lnavgsalary_cpi,
  treatment = anzia2012$oncycle,
  post_treatment = c(2007, 2008, 2009),
  id_subject = anzia2012$district,
  id_time = anzia2012$year,
  ylim = c(10.5, 10.8)
)
}
