% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{did_plot}
\alias{did_plot}
\title{DiD plot}
\usage{
did_plot(formula, data, post_treatment, id_subject = NULL, id_time,
  diff_order = 0, alpha = 0.05, ci = TRUE, xlim = NULL,
  ylim = NULL, col = NULL, loc = NULL, lwd = NULL, ...)
}
\arguments{
\item{formula}{\code{outcome ~ treatment}.}

\item{data}{data set}

\item{post_treatment}{a vector of time index indicating post treatment periods.}

\item{id_subject}{a variable name of unit index.
This should be left as \code{NULL} when repeated cross-section data is used.}

\item{id_time}{a variable name of time index.}

\item{diff_order}{Order of differences. \code{diff_order = 0} generates a plot based on the original outcome.
\code{diff_order = 1} generates a plot based on differenced outcomes.}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{col}{col}

\item{loc}{Location of legend. See \code{\link{legend}}.}

\item{lwd}{Line width.}
}
\description{
Generate a plot for difference-in-differences design.
}
\examples{
# load package
require(DIDdesign)

# load data
data(anzia2012)

# original series
did_plot(lnavgsalary_cpi ~ oncycle, data = anzia2012,
  post_treatment = c(2007, 2008, 2009),
  id_subject = "district",
  id_time = "year",
  diff_order = 0,
  main = 'Original Series'
)

# differenced series
did_plot(lnavgsalary_cpi ~ oncycle, data = anzia2012,
  post_treatment = c(2007, 2008, 2009),
  id_subject = "district",
  id_time = "year",
  ylim = c(-0.05, 0.05),
  diff_order = 1,
  main = 'Differenced Series'
)
}
\seealso{
Other main functions: \code{\link{did_plot_selection}},
  \code{\link{did}}, \code{\link{plot.diddesign}},
  \code{\link{summary.diddesign}}
}
\concept{main functions}
