% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{did_plot_selection}
\alias{did_plot_selection}
\title{Plot function for selection}
\usage{
did_plot_selection(data, alpha = 0.05, equivalence = TRUE, eL = NULL,
  eU = NULL, xlim = NULL, ylim = NULL, loc = NULL, ...)
}
\arguments{
\item{data}{\code{diddesign} object.
Typically, it is an output from \code{\link{did}} function.}

\item{alpha}{Level of rejection.
Confidence intervals are constructed with the 1-alpha and 1 - alpha/2 levels.}

\item{equivalence}{boolean; if \code{TRUE}, equivalence region is superimposed on a plot.
See also \code{eL} and \code{eU} arguments to set the upper and the lower bound of a region.}

\item{eL}{lower bound of equivalence region. If unspecified, the smallest value of 1 - alpha level sets is used.}

\item{eU}{upper bound of equivalence region. If unspecified, the largest value of 1 - alpha level sets is used.}

\item{xlim}{xlim. See \code{\link{plot}} function.}

\item{ylim}{ylim. See \code{\link{plot}} function.}

\item{loc}{location of legend. See \code{link{legend}} function.}
}
\description{
Plot function for selection
}
\examples{
# load package
require(DIDdesign)

# load data
data(anzia2012)

# fit models
fit1 <- did(lnavgsalary_cpi ~ oncycle, data = anzia2012,
            id_subject = "district", id_time = "year",
            post_treatment = c(2007, 2008, 2009),
            method = "parametric")
fit2 <- did(lnavgsalary_cpi ~ oncycle, data = anzia2012,
            id_subject = "district", id_time = "year",
            post_treatment = c(2007, 2008, 2009),
            method = "nonparametric", se_boot = FALSE)

# make a plot with equivalence region
par(mfrow = c(1,2), mar = c(4, 2.5, 3.5, 1))
did_plot_selection(fit1, equivalence = TRUE, eL = -0.005, eU = 0.005,
 ylim = c(-0.015, 0.015), main = "Parametric")
did_plot_selection(fit2, equivalence = TRUE, eL = -0.005, eU = 0.005,
 ylim = c(-0.015, 0.015), main = "Nonparametric")
}
