% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-check-sad.R
\name{did_sad_placebo}
\alias{did_sad_placebo}
\title{SA Placebo Regression}
\usage{
did_sad_placebo(formula, dat_panel, treatment, outcome, option, show_warning)
}
\arguments{
\item{formula}{A formula.}

\item{dat_panel}{A panel data.}

\item{treatment}{Name of the treatment variable.}

\item{outcome}{Name of the outcome variable.}

\item{option}{A list of options.}
}
\value{
A matrix of placebo estimates.
}
\description{
SA Placebo Regression
}
\keyword{internal}
