% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIDdesign.R
\docType{package}
\name{DIDdesign-package}
\alias{DIDdesign}
\alias{DIDdesign-package}
\title{Difference-in-Differences Design}
\description{
This packages provides tools for estimating treatment effects
 and diagnosing difference-in-differences designs.
}
\details{
The package has the following main functions:
\itemize{
  \item \code{\link{did_plot}}: create a DiD plot to visualize difference-in-differences design.
  \item \code{\link{did}}: estimate treatment effects.
  \item \code{\link{plot.diddesign}}: visualize estimates.
  \item \code{\link{summary.diddesign}}: summarize estimates.
  \item \code{\link{did_plot_selection}}: visualize statistics used for model selection.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/naoki-egami/DIDdesign}
  \item Report bugs at \url{https://github.com/naoki-egami/DIDdesign/issues}
}

}
\author{
\strong{Maintainer}: Soichiro Yamauchi \email{email@gmail.com} (\href{https://orcid.org/0000-0003-0554-2717}{ORCID})

Authors:
\itemize{
  \item Naoki Egami \email{email@gmail.com} (\href{https://orcid.org/0000-0002-5491-2174}{ORCID})
}

}
