% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIDdesign.R
\docType{package}
\name{DIDdesign-package}
\alias{DIDdesign}
\alias{DIDdesign-package}
\title{Difference-in-Differences Design}
\description{
This packages provides tools for estimating treatment effects 
 and diagnosing difference-in-differences designs.
}
\details{
The package has two main functions:
\itemize{
  \item \code{\link{did_data}}: format data for analysis
  \item \code{\link{did}}: estimate treatment effects 
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/naoki-egami/DIDdesign}
  \item Report bugs at \url{https://github.com/naoki-egami/DIDdesign/issues}
}

}
\author{
\strong{Maintainer}: Soichiro Yamauchi \email{email@gmail.com} (0000-0003-0554-2717)

Authors:
\itemize{
  \item Naoki Egami \email{email@gmail.com}
}

}
