% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmmT.R
\name{didgmmT}
\alias{didgmmT}
\title{DID-GMM with mth order parallel treand assumption}
\usage{
didgmmT(Y, D, M, ep = 0.01, max_trial = 100, only_beta = FALSE,
  only_oneM = FALSE)
}
\arguments{
\item{Y}{Outcome matrix.}

\item{D}{Treatment vector.}

\item{M}{Order specification.}

\item{ep}{small constant added to diagonal elements of the weighting matrix in GMM. Default is 0.01.}

\item{only_beta}{A boolean argument.
If \code{TRUE} ATT is estimated only using moment conditions related to the last time-period.
Default is \code{FALSE}.}
}
\description{
DID-GMM with mth order parallel treand assumption
}
