% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didgmm-selection.R
\name{fe_selection}
\alias{fe_selection}
\title{Model selectin based on linear FE model}
\usage{
fe_selection(data, fm_list, post_time, alpha = 0.05)
}
\description{
Model selectin based on linear FE model
}
\keyword{internal}
