% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{generate_tab_parametric}
\alias{generate_tab_parametric}
\title{Generate Table for Parametric Method}
\usage{
generate_tab_parametric(obj, full = FALSE)
}
\arguments{
\item{obj}{an object of \code{diddesign} class.}

\item{full}{a boolean; if \code{TRUE} all results are attached.}
}
\description{
Generate Table for Parametric Method
}
\keyword{internal}
