% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{plot.diddesign}
\alias{plot.diddesign}
\title{Plot function}
\usage{
\method{plot}{diddesign}(data, xlim = NULL, ylim = NULL, col = NULL,
  lwd = NULL, full = FALSE, selection = TRUE, ...)
}
\arguments{
\item{data}{Input object. This should be either diddesign object or diddesign_data object.
If diddesign object is provided, ATT estimates with 95% confidence intervals are plotted.
If did_double.data objec is provided, mean outcomes for the treated and the control are plotted.}

\item{xlim}{xlim in plot function. If left \code{NULL}, it's automatically set.}

\item{ylim}{ylim in plot function. If left \code{NULL}, it's automatically set.}

\item{full}{A boolean argument, if \code{TRUE} ATT estimates for all models are plotted,
otherwise only estimates from the selected model will be shown. Default is \code{FALSE}.}

\item{selection}{A boolean. If \code{TRUE} statistics used for model selection is displayed next to the effect plot.}

\item{...}{additional arguments supplied to the plot function.}
}
\description{
Plot function
}
\seealso{
Other main functions: \code{\link{did_plot_selection}},
  \code{\link{did_plot}}, \code{\link{did}},
  \code{\link{summary.diddesign}}
}
\concept{main functions}
