% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-plot.R
\name{plot_diddesign_data}
\alias{plot_diddesign_data}
\title{Baseline plot function for did_plot}
\usage{
plot_diddesign_data(
  data,
  panel = TRUE,
  diff_order = 0,
  alpha = 0.05,
  ci = TRUE,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  loc = NULL,
  lwd = NULL,
  ...
)
}
\arguments{
\item{data}{Two possible inputs. If \code{panel = TRUE}, input should be a \code{diddesign_data} class object.
If \code{panel = FALSE}, it should be a list consists of \code{"y1mean"}, \code{"y0mean"} and \code{"id_time"}.}

\item{panel}{A boolean argument. \code{TRUE} if data are in the panel format and \code{FALSE} if repeated cross-section.}

\item{diff_order}{Either \code{0} (original series) or \code{1} (differenced series).}
}
\description{
Baseline plot function for did_plot
}
\keyword{internal}
