% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{print.summary.DIDdesign}
\alias{print.summary.DIDdesign}
\title{Print}
\usage{
\method{print}{summary.DIDdesign}(x, ...)
}
\arguments{
\item{x}{An object of \code{summary.DIDdesign} class. This is typically an output of \code{summary.DIDdesign()} function.}
}
\description{
Print
}
