% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad.R
\name{sa_double_did}
\alias{sa_double_did}
\title{Compute the time-weighted DID and sDID}
\usage{
sa_double_did(formula, dat_panel, treatment, outcome, option)
}
\value{
A vector of two elements: Time-weighted DID (the first element) and time-weighted sequential DID (second).
}
\description{
Compute the time-weighted DID and sDID
}
\keyword{internal}
