% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-sad-simulate.R
\name{simulate_sad}
\alias{simulate_sad}
\title{Generate staggered adoption design}
\usage{
simulate_sad(n_obs, n_times)
}
\arguments{
\item{n_obs}{Number of units.}

\item{n_time}{Number of time periods.}
}
\description{
Generate staggered adoption design
}
\keyword{internal}
