% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did-summary.R
\name{summary.diddesign}
\alias{summary.diddesign}
\title{Summarize Output from DiD Function}
\usage{
\method{summary}{diddesign}(obj, full = TRUE)
}
\arguments{
\item{obj}{an object of \code{diddesign_data} or \code{diddesign} class.}
}
\description{
Summarize Output from DiD Function
}
